// web/src/Definitions/Enums/projectProposalStage.enum.ts
export enum ProjectProposalStage {
  PENDING = 'PENDING',
  REJECTED = 'REJECTED',
  APPROVED = 'APPROVED',
  PDD_SUBMITTED = 'PDD_SUBMITTED',
  PDD_REJECTED_BY_CERTIFIER = 'PDD_REJECTED_BY_CERTIFIER',
  PDD_APPROVED_BY_CERTIFIER = 'PDD_APPROVED_BY_CERTIFIER',
  PDD_REJECTED_BY_DNA = 'PDD_REJECTED_BY_DNA',
  PDD_APPROVED_BY_DNA = 'PDD_APPROVED_BY_DNA',
  VALIDATION_REPORT_SUBMITTED = 'VALIDATION_REPORT_SUBMITTED',
  VALIDATION_REPORT_REJECTED = 'VALIDATION_REPORT_REJECTED',
  AUTHORISED = 'AUTHORISED',
}